Memory usage tweaks
