+++ build.sbt
@@ -3,7 +3,7 @@ ThisBuild / organization := "io.appthreat"
 ThisBuild / version      := "2.5.16"
 ThisBuild / scalaVersion := "3.6.2"
 
-val cpgVersion = "2.1.5"
+val cpgVersion = "2.1.6"
 
 lazy val platform          = Projects.platform
 lazy val console           = Projects.console
