+++ platform/frontends/c2cpg/src/main/scala/io/appthreat/c2cpg/passes/AstCreationPass.scala
@@ -3,9 +3,9 @@ package io.appthreat.c2cpg.passes
 import io.appthreat.c2cpg.Config
 import io.appthreat.c2cpg.astcreation.AstCreator
 import io.appthreat.c2cpg.parser.{CdtParser, FileDefaults, HeaderFileFinder}
-import io.appthreat.x2cpg.{Ast, AstEdge, SourceFiles, ValidationMode}
+import io.appthreat.x2cpg.{Ast, AstCache, SourceFiles, ValidationMode}
+import io.appthreat.x2cpg.AstCache.AstBitcode
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.codepropertygraph.generated.nodes.NewNode
 import io.shiftleft.passes.StreamingCpgPass
 import upickle.default.*
 
@@ -14,143 +14,12 @@ import java.security.MessageDigest
 import java.util.concurrent.*
 import java.util.regex.Pattern
 import scala.concurrent.duration.*
-import scala.jdk.CollectionConverters.*
 import scala.util.{Failure, Success, Try}
 import scala.util.matching.Regex
 
 object AstCreationPass:
   private val theoriticalMaxProcs  = Math.max(1, Runtime.getRuntime.availableProcessors() / 3)
   private val maxConcurrentParsers = Math.min(4, theoriticalMaxProcs)
-  private val astCreationSemaphore = new Semaphore(maxConcurrentParsers)
-  private val parseExecutor        = Executors.newCachedThreadPool()
-
-  case class CachedNode(className: String, properties: Map[String, ujson.Value])
-      derives ReadWriter
-
-  case class CachedEdge(srcId: Int, dstId: Int, label: String)
-      derives ReadWriter
-
-  case class CachedAst(
-    rootIdx: Option[Int],
-    nodes: List[CachedNode],
-    edges: List[CachedEdge]
-  ) derives ReadWriter
-
-  object Serialization:
-
-    def toCached(ast: Ast): CachedAst =
-      val allNodes    = ast.nodes.distinct.toVector
-      val nodeToIndex = allNodes.zipWithIndex.toMap
-
-      val cachedNodes = allNodes.map { n =>
-        val propsMap = n.properties match
-          case m: java.util.Map[?, ?]        => m.asScala
-          case m: scala.collection.Map[?, ?] => m
-          case null                          => Map.empty
-
-        val safeProps = propsMap.map {
-            case (k: String, v) => k          -> toUjson(v)
-            case (k, v)         => k.toString -> toUjson(v)
-        }.toMap
-
-        CachedNode(n.getClass.getName, safeProps)
-      }.toList
-
-      def serializeEdgeList(edges: scala.collection.Seq[AstEdge], label: String): List[CachedEdge] =
-          edges.collect {
-              case e if nodeToIndex.contains(e.src) && nodeToIndex.contains(e.dst) =>
-                  CachedEdge(nodeToIndex(e.src), nodeToIndex(e.dst), label)
-          }.toList
-
-      val allEdges =
-          serializeEdgeList(ast.edges, "AST") ++
-              serializeEdgeList(ast.refEdges, "REF") ++
-              serializeEdgeList(ast.bindsEdges, "BINDS") ++
-              serializeEdgeList(ast.receiverEdges, "RECEIVER") ++
-              serializeEdgeList(ast.argEdges, "ARGUMENT") ++
-              serializeEdgeList(ast.conditionEdges, "CONDITION") ++
-              serializeEdgeList(ast.captureEdges, "CAPTURE")
-
-      val rootIdx = ast.root.flatMap(nodeToIndex.get)
-
-      CachedAst(rootIdx, cachedNodes, allEdges)
-    end toCached
-
-    def fromCached(cAst: CachedAst)(implicit validation: ValidationMode): Ast =
-      val nodesVector = cAst.nodes.map(reconstructNode).toVector
-
-      def getEdges(label: String): Seq[AstEdge] =
-          cAst.edges.collect {
-              case ce if ce.label == label =>
-                  AstEdge(nodesVector(ce.srcId), nodesVector(ce.dstId))
-          }
-
-      Ast(
-        nodesVector,
-        getEdges("AST"),
-        getEdges("CONDITION"),
-        getEdges("REF"),
-        getEdges("BINDS"),
-        getEdges("RECEIVER"),
-        getEdges("ARGUMENT"),
-        getEdges("CAPTURE")
-      )
-
-    private def toUjson(v: Any): ujson.Value = v match
-      case s: String       => ujson.Str(s)
-      case b: Boolean      => ujson.Bool(b)
-      case i: Int          => ujson.Num(i)
-      case l: Long         => ujson.Num(l.toDouble)
-      case d: Double       => ujson.Num(d)
-      case xs: Iterable[?] => ujson.Arr.from(xs.map(toUjson))
-      case xs: Array[?]    => ujson.Arr.from(xs.map(toUjson))
-      case other           => ujson.Str(other.toString)
-
-    private def fromUjson(v: ujson.Value): Any = v match
-      case ujson.Str(s)  => s
-      case ujson.Bool(b) => b
-      case ujson.Num(n) =>
-          if n.isValidInt then n.toInt
-          else if n.toLong.toDouble == n then n.toLong
-          else n
-      case ujson.Arr(arr) => arr.map(fromUjson).toList
-      case _              => v.toString
-
-    private def reconstructNode(cn: CachedNode): NewNode =
-      if !cn.className.startsWith("io.shiftleft.codepropertygraph.generated.nodes.") then
-        throw new SecurityException(s"Illegal class in cache: ${cn.className}")
-
-      val clazz = Class.forName(cn.className)
-      val node  = clazz.getDeclaredConstructor().newInstance().asInstanceOf[NewNode]
-
-      cn.properties.foreach { case (k, v) =>
-          try
-            val camelName = snakeToCamel(k)
-            val methods   = clazz.getMethods
-            var method    = methods.find(_.getName == camelName)
-            if method.isEmpty then
-              method = methods.find(_.getName.equalsIgnoreCase(k.replace("_", "")))
-
-            if method.isDefined && method.get.getParameterCount == 1 then
-              val arg = fromUjson(v)
-              method.get.invoke(node, arg.asInstanceOf[Object])
-          catch
-            case _: Exception =>
-      }
-      node
-    end reconstructNode
-
-    private def snakeToCamel(s: String): String =
-      val tokens = s.split("_")
-      if tokens.isEmpty then return s.toLowerCase
-      val sb = new StringBuilder(tokens(0).toLowerCase)
-      for i <- 1 until tokens.length do
-        val t = tokens(i).toLowerCase
-        if t.nonEmpty then
-          sb.append(t.substring(0, 1).toUpperCase).append(t.substring(1))
-      sb.toString
-  end Serialization
-end AstCreationPass
 
 class AstCreationPass(
   cpg: Cpg,
@@ -160,6 +29,8 @@ class AstCreationPass(
 ) extends StreamingCpgPass[String](cpg):
 
   import AstCreationPass.*
+
+  private val parseExecutor = Executors.newFixedThreadPool(maxConcurrentParsers)
 
   private val sharedHeaderFileFinder = new HeaderFileFinder(config.inputPath)
   private val EscapedFileSeparator   = Pattern.quote(java.io.File.separator)
@@ -174,6 +45,14 @@ class AstCreationPass(
       case Failure(e) =>
           println(s"Warning: Failed to create cache directory ${config.cacheDir}: ${e.getMessage}")
       case Success(_) =>
+
+  override def finish(): Unit =
+      try
+        parseExecutor.shutdownNow()
+      catch
+        case e: Exception => println(s"Error shutting down parse executor: ${e.getMessage}")
+      finally
+        super.finish()
 
   override def generateParts(): Array[String] =
       SourceFiles
@@ -193,7 +72,6 @@ class AstCreationPass(
     val file2OffsetTable = new ConcurrentHashMap[String, Array[Int]]()
     val fileHash         = if config.enableAstCache then Option(computeEntryHash(path)) else None
 
-    AstCreationPass.astCreationSemaphore.acquire()
     try
       val cachedAst: Option[Ast] = fileHash.flatMap(h => checkAndLoadCache(h, config.cacheDir))
       cachedAst match
@@ -212,9 +90,7 @@ class AstCreationPass(
                           translationUnit,
                           file2OffsetTable,
                           fileHash
-                        )(
-                          using config.schemaValidation
-                        ).createAst(),
+                        )(using config.schemaValidation).createAst(),
                     timeoutDuration
                   )
                   if !config.onlyAstCache then diffGraph.absorb(localDiff)
@@ -228,13 +104,8 @@ class AstCreationPass(
           println(
             s"Exception processing file $path: ${cause.getClass.getSimpleName} - ${cause.getMessage}"
           )
-//          cause.printStackTrace()
       case e: Throwable =>
           println(s"Exception processing file $path: ${e.getClass.getSimpleName} - ${e.getMessage}")
-//          e.printStackTrace()
-    finally
-      AstCreationPass.astCreationSemaphore.release()
-      sharedHeaderFileFinder.clear()
     end try
   end runOnPart
 
@@ -245,14 +116,14 @@ class AstCreationPass(
     digest.digest().map("%02x".format(_)).mkString
 
   private def checkAndLoadCache(hash: String, cacheDir: String): Option[Ast] =
-    val cacheFile = Paths.get(cacheDir, s"$hash.json")
+    val cacheFile = Paths.get(cacheDir, s"$hash.ast")
     if !Files.exists(cacheFile) then return None
 
     try
       val bytes                            = Files.readAllBytes(cacheFile)
-      val cachedAst                        = readBinary[CachedAst](bytes)
+      val bitcode                          = readBinary[AstBitcode](bytes)
       implicit val valMode: ValidationMode = config.schemaValidation
-      Some(Serialization.fromCached(cachedAst))
+      Some(AstCache.fromBitcode(bitcode))
     catch
       case e: Exception =>
           println(s"Failed to load cache for $hash: ${e.getClass.getSimpleName} - ${e.getMessage}")
@@ -260,7 +131,7 @@ class AstCreationPass(
           None
 
   private def runWithTimeout[T](block: () => T, timeout: FiniteDuration): T =
-    val future = AstCreationPass.parseExecutor.submit(new Callable[T]:
+    val future = parseExecutor.submit(new Callable[T]:
       override def call(): T = block()
     )
     try
@@ -272,6 +143,8 @@ class AstCreationPass(
       case e: InterruptedException =>
           future.cancel(true)
           val cause = e.getCause
-          cause.printStackTrace()
           throw new InterruptedException(s"Operation interrupted - ${cause.getMessage}")
+      case e =>
+          future.cancel(true)
+          throw e
 end AstCreationPass
